funk: move insert_para API to program cache
This patch moves a app-specific API out of funk to its user
