+++ src/flamenco/runtime/tests/fd_dump_pb.c
@@ -3,8 +3,8 @@
 #include "generated/invoke.pb.h"
 #include "generated/txn.pb.h"
 #include "generated/vm.pb.h"
+#include "../context/fd_exec_slot_ctx.h"
 #include "../fd_system_ids.h"
-#include "../fd_runtime.h"
 #include "../program/fd_address_lookup_table_program.h"
 #include "../../../ballet/nanopb/pb_encode.h"
 #include "../program/fd_program_cache.h"
