+++ src/funk/Local.mk
@@ -7,8 +7,6 @@ $(call make-unit-test,test_funk,test_funk,fd_funk fd_util)
 $(call run-unit-test,test_funk)
 $(call make-unit-test,test_funk_concur,test_funk_concur,fd_funk fd_util)
 #$(call run-unit-test,test_funk_concur,)
-$(call make-unit-test,test_funk_concur2,test_funk_concur2,fd_funk fd_util)
-#$(call run-unit-test,test_funk_concur2,)
 $(call make-unit-test,test_funk_rec,test_funk_rec test_funk_common,fd_funk fd_util)
 $(call run-unit-test,test_funk_rec)
 $(call make-unit-test,test_funk_txn,test_funk_txn test_funk_common,fd_funk fd_util)
